Triangle Coordinates Code Written In C#
The task was to calculate the triangle coordinates for an image with right triangles such that for a given row (A-F) and column (1-12) you can produce the layout: 60 x 60 Pixels. Each non-hypotenuse side of the triangle is 10 pixels. Lastly, given the vertex coordinates, to calculate the row and column for the triangle.